Audio news

+++ app/src/main/java/com/hon/librarytest02/util/Util.java
@@ -26,7 +26,7 @@ public class Util {
         return LibraryTest.sContext.getResources().getString(resId);
     }
 
-    public static int dp2px(float value) {
+    public static int dp2px(int value) {
         final float scale = Resources.getSystem().getDisplayMetrics().densityDpi;
         return (int) (value * (scale / 160) + 0.5f);
     }
